Move fixtures to where they're needed
This avoids an import error due to `PROJECT_CONFIGS_DIR` not being defined in `conftest` until after the imports
